assignment2/requirements.txt
haversine==2.7.0
mysql-connector-python==8.0.30
tabulate==0.8.9
python-decouple==3.6
python-dotenv
pandas
